Moved some  time slots from Store reducer to Delivery reducer
