+++ src/redux/Delivery/selectors.js
@@ -1,1 +1,2 @@
 export const selectStore = state => state.delivery.store;
+export const selectTimeSlots = state => state.delivery.timeSlots;
