+++ src/redux/Store/selectors.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 import { createSelector } from 'reselect';
+import { selectTimeSlots } from '../Delivery/selectors';
 
 export const selectStore = state => state.store.store;
 export const selectDeliveries = state => state.store.deliveries;
-export const selectTimeSlots = state => state.store.timeSlots;
 
 export const selectTimeSlot = createSelector(
   selectStore,
